Move GameObjects type into own folder
