+++ darwin-types/State.ts
@@ -1,4 +1,4 @@
-import { GameObject, ObjectId } from './GameObject';
+import { GameObject, ObjectId } from './game-objects/GameObject';
 
 export interface State {
   objectMap: Record<ObjectId, GameObject>;
