+++ darwin-types/game-objects/GameObject.ts
@@ -1,8 +1,9 @@
-import Position from './Position';
+import Position from '../Position';
 
 export type ObjectId = string;
 
 export interface GameObject {
   id: ObjectId;
+  type: string;
   position: Position;
 }
